Add support for fully skipping tests
Some tests are incorrect and known to crash even on correct JDWP
implementations. This can lead to error logs being hard to parse since
there will be spurious crashes in them. To improve this situation we
add the ability to entirely suppress the running of specific tests by
passing them, in a comma-delimited list, with the
jpda.settings.badTestCases variable.

Test: ./art/tools/run-libjdwp-tests.sh --mode=host --skip-test 'org.apache.harmony.jpda.tests.jdwp.VirtualMachine_VersionTest#testVersion001'
      Examine output
Bug: 141907697

Change-Id: Icbac5f3cd82689e04660448fd2796159c5e11da2