Deflake SimpleFormatterTest.
This test was passing on UTC, Europe/London and
America/Los_Angeles, but failing on Nuuk or America/Santiago.

Bug was in current year calculation: LogRecord's time is set to 0,
but Calendar was set to 12321312 millis, which about 3.5 hours. So,
if time zone's offset is negative and small enough 0 and 12321312
epoch millis will be in different years - in 1969 and 1970
correspondingly.

Bug: 155075213
Fix: 155075213

Test: atest CtsLibcoreTestCases:org.apache.harmony.logging.tests.java.util.logging.SimpleFormatterTest
Test: passes will all the time zones mentioned above

Change-Id: I19b214c5a809d084e5ed1b67aca96f9c44b2519b